feat(*):code splitting and modularity
