+++ web/src/views/Home.vue
@@ -1,6 +1,6 @@
 <script setup>
-import PcHome from "./PcHome.vue";
-import MobileHome from "./MobileHome.vue";
+import PcHome from "@/views/PcHome/PcHome.vue";
+import MobileHome from "@/views/MobileHome/MobileHome.vue";
 import pageStore from "@/stores/model/page";
 
 const pageWidth = computed(() => pageStore().getScreenWidth());
