+++ web/vite.config.js
@@ -45,7 +45,8 @@ export default defineConfig({
   server: {
     proxy: {
       "/api": {
-        target: "http://127.0.0.1:8080",
+        // target: "http://127.0.0.1:8080",
+        target: "http://175.155.64.58:8080",
         changeOrigin: true,
         // rewrite: (path) => path.replace(/^\/api/, ""),
       },
